Propose adding a License
Hey, I'm a mentor of FRC team 4909. I'm hoping to use your implementation as the basis for a project we are working on. If you wouldn't mind adding a license we would very much appreciate your time in doing so. This pull request suggests the MIT license to allow for maximum community flexibility.